demo/requirements.txt
# Hugging Face Space Requirements for VideoMaMa Demo

# Core frameworks
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.30.0

# Gradio for UI
gradio==5.12.0

# Image and video processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# SAM2 dependencies
git+https://github.com/facebookresearch/sam2.git

# Additional utilities
accelerate>=0.20.0
einops>=0.6.0
tqdm>=4.65.0
safetensors>=0.3.0

# For video export
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
pydantic==2.10.6